feat: W3 桌機 TitleBar action 恢復 icon + 可見 label（推翻 v2.31.90）(v2.57.27)
owner 2026-07-24 裁定改回 HIG header 慣例：桌機動作鈕顯示 icon + 文字（可辨識、
無障礙），不再 icon-only。.tp-titlebar-action-label top-level display:none → inline
（桌機顯示），只在 @media (max-width:760px) 隱藏（手機 icon-only 圓形）；icon-only
變體顯式保持不顯 label。title tooltip 保留。lock 測試改寫為 w3-titlebar-action-labels。
真瀏覽器驗：桌機收藏頁「探索」顯示羅盤 icon + 文字。全套 470 files/4090 tests 綠。
