+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "tripline",
-  "version": "2.57.26",
+  "version": "2.57.27",
   "private": true,
   "scripts": {
     "dev": "concurrently -n vite,api -c cyan,green \"vite\" \"wrangler pages dev dist --local --port 8788 --d1 DB\"",
